+++ backends/backend-robovm/src/com/badlogic/gdx/backends/iosrobovm/IOSSound.java
@@ -32,16 +32,16 @@ public class IOSSound implements Sound{
     }
 
     @Override
-    public long play(float volume){
+    public int play(float volume){
         return play(volume, 1, 0, false);
     }
 
     @Override
-    public long play(float volume, float pitch, float pan){
+    public int play(float volume, float pitch, float pan){
         return play(volume, pitch, pan, false);
     }
 
-    public long play(float volume, float pitch, float pan, boolean loop){
+    public int play(float volume, float pitch, float pan, boolean loop){
         if(streamIds.size == 8) streamIds.pop();
         ALSource soundSource = OALSimpleAudio.sharedInstance().playBuffer(soundBuffer, volume, pitch, pan, loop);
         if(soundSource == null) return -1;
@@ -51,17 +51,17 @@ public class IOSSound implements Sound{
     }
 
     @Override
-    public long loop(){
+    public int loop(){
         return play(1, 1, 0, true);
     }
 
     @Override
-    public long loop(float volume){
+    public int loop(float volume){
         return play(volume, 1, 0, true);
     }
 
     @Override
-    public long loop(float volume, float pitch, float pan){
+    public int loop(float volume, float pitch, float pan){
         return play(volume, pitch, pan, true);
     }
 
@@ -80,31 +80,31 @@ public class IOSSound implements Sound{
     }
 
     @Override
-    public void stop(long soundId){
+    public void stop(int soundId){
         ALSource source;
         if((source = getSoundSource(soundId)) != null) source.stop();
     }
 
     @Override
-    public void setLooping(long soundId, boolean looping){
+    public void setLooping(int soundId, boolean looping){
         ALSource source;
         if((source = getSoundSource(soundId)) != null) source.setLooping(looping);
     }
 
     @Override
-    public void setPitch(long soundId, float pitch){
+    public void setPitch(int soundId, float pitch){
         ALSource source;
         if((source = getSoundSource(soundId)) != null) source.setPitch(pitch);
     }
 
     @Override
-    public void setVolume(long soundId, float volume){
+    public void setVolume(int soundId, float volume){
         ALSource source;
         if((source = getSoundSource(soundId)) != null) source.setVolume(volume);
     }
 
     @Override
-    public void setPan(long soundId, float pan, float volume){
+    public void setPan(int soundId, float pan, float volume){
         ALSource source;
         if((source = getSoundSource(soundId)) != null){
             source.setPan(pan);
@@ -129,13 +129,13 @@ public class IOSSound implements Sound{
     }
 
     @Override
-    public void pause(long soundId){
+    public void pause(int soundId){
         ALSource source;
         if((source = getSoundSource(soundId)) != null) source.setPaused(true);
     }
 
     @Override
-    public void resume(long soundId){
+    public void resume(int soundId){
         ALSource source;
         if((source = getSoundSource(soundId)) != null) source.setPaused(false);
     }
